requirements.txt
discord.py[speed]
python-dotenv
